Update compression level config option naming.
The values for the config option now align to the naming used by the Google Closure Compiler java library. The name of the config option itself follows better the naming convention used by other plugins that build on top of the Resources Plugin such as the cached-resources plugin and the lesscss-resources plugin.